Add setter for internal property and refine extensions
