--- eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Extensions/StringExtensions.cs
@@ -1,25 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using Humanizer;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Azure.Generator.Management.Extensions
-{
-    internal static class StringExtensions
-    {
-        public static IEnumerable<string> SplitByCamelCase(this string camelCase)
-        {
-            return camelCase.Humanize().Split(' ').Select(w => w.FirstCharToUpperCase());
-        }
-
-        public static string FirstCharToUpperCase(this string str)
-        {
-            if (string.IsNullOrEmpty(str) || char.IsUpper(str[0]))
-                return str;
-
-            return char.ToUpper(str[0]) + str.Substring(1);
-        }
-    }
-}
